requirements.txt
gradio>=4.0.0
sentence-transformers>=2.2.0
pinecone-client>=3.0.0
PyMuPDF>=1.23.0
rank-bm25>=0.2.2
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.36.0
tqdm>=4.65.0
